refactor: clean up file structure and minor style changes
